panel/requirements.txt
Adafruit-BBIO>=1.2.0
python-engineio>=4.3.4
python-socketio>=5.8.0
adafruit_mcp4728>=1.3.5
Adafruit-Blinka>=8.18.1
